C++
Please complete the following two coding assignments. 

Weight Conversion problem - Write a program the prompts the user to enter the weight of a person in kilograms and output the equivalent weight in pounds. Output both the weights rounded to two decimal places (note that 1 kilogram equals 2.2 lbs.). format your output with two decimal places.  Save this problem as CIT-100-<Section>_M4_<LName><FName>WeightConversion.cpp (no spaces).
 

Logic Errors -During each summer, John and Jessica grow vegetables in their backyard and buy seeds and fertilizer from a local Nursery. The nursery carries different types of vegetable fertilizers in various bag sizes. When buying a particular fertilizer, they want to know the price of the fertilizer per pound and the cost of fertilizing per square foot. The following program prompts the user to enter the size of the fertilizer bag, in pounds, the cost of the bag, and the area, in square feet, that can be covered by the bag. The program should output the desired result. However, the program contains logic errors. Find and correct the logic errors so that the program works properly. The program is pre-written and is attached.